requirements.txt
decord==0.6.0
einops==0.3.0
imageio==2.9.0
numpy==1.24.2
omegaconf==2.1.1
opencv_python
pandas==2.0.0
Pillow==9.5.0
pytorch_lightning==1.8.3
PyYAML==6.0
setuptools==65.6.3
fairscale
torch==2.2.1
torchvision==1.7.1
tqdm==4.65.0
transformers==4.25.1
moviepy
av
xformers==0.0.25
flash_attn
gradio
timm
scikit-learn 
open_clip_torch
kornia
wandb
